requirements.txt
pandas
openpyxl
requests
beautifulsoup4
openai  
python-dotenv
pylint 
black
pytest
faiss-cpu
rank_bm25 
sentence-transformers
selenium
webdriver-manager
scikit-learn
bert-score          
rouge-score         
ragas               
deepeval            
transformers        
torch               
datasets            
nltk
langchain-openai
ragas.llms